Split the interpreter into per-command handler modules
The Interpreter class dispatched commands via a string index signature
([key: string]: Method), which forced @ts-ignore on every counter field
and allowed dispatch to any method on the class. Each command now lives
in its own module under src/interpreter/commands/, wired up through an
explicit registry map; the eight copy-pasted tool-selection methods
collapse into a single handler that reads the tool number from the
gcode. Behavior is unchanged and the import path stays ./interpreter.
