--- src/interpreter.ts
@@ -1,421 +1,0 @@
-import { Path, PathType } from './path';
-import { GCodeCommand } from './parser/gcode-parser';
-import { Job } from './job';
-import { State } from './state';
-
-type Method = (..._args: unknown[]) => unknown;
-
-/**
- * Resolves a state's position for rendering.
- * @param state - Current job state
- * @returns The position as concrete `x`, `y`, `z` numbers
- * @remarks
- * An axis that has not been homed has an unknown (`undefined`) position. The
- * interpreter chooses to assume the origin (`0`) for such axes so the viewer can
- * still render best-effort (see #361); `state.isHomed` lets a consumer tell these
- * assumed coordinates from real ones.
- */
-function resolvePosition(state: State): { x: number; y: number; z: number } {
-  return { x: state.x ?? 0, y: state.y ?? 0, z: state.z ?? 0 };
-}
-
-/**
- * Interprets and executes G-code commands, updating the job state accordingly
- *
- * @remarks
- * This class handles the execution of G-code commands, translating them into
- * movements and state changes in the print job. It supports common G-code commands
- * including linear moves (G0/G1), arcs (G2/G3), unit changes (G20/G21), and tool selection.
- */
-export class Interpreter {
-  [key: string]: Method;
-
-  // TODO: maybe these props should move to the Job class
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  public retractions = 0;
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  public deretractions = 0;
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  public feedrateChanges = 0;
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  public others = 0;
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  public points = 0; // for reference, how many points were added to the job
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  public extrusionDistance = 0;
-
-  /**
-   * Executes an array of G-code commands, updating the provided job
-   * @param commands - Array of GCodeCommand objects to execute
-   * @param job - Job instance to update (default: new Job)
-   * @returns The updated job instance
-   */
-  execute(commands: GCodeCommand[], job = new Job()): Job {
-    job.resumeLastPath();
-    commands.forEach((command) => {
-      if (command.gcode !== undefined) {
-        if (typeof this[command.gcode] !== 'function') {
-          return;
-        }
-        const method = this[command.gcode] as Method;
-        method.bind(this)(command, job);
-      }
-    });
-    job.finishPath();
-
-    return job;
-  }
-
-  /**
-   * Executes a linear move command (G0/G1)
-   * @param command - GCodeCommand containing move parameters
-   * @param job - Job instance to update
-   * @remarks
-   * Handles both rapid moves (G0) and linear moves (G1). Updates the job state
-   * and adds points to the current path based on the command parameters.
-   * G0 is for rapid moves (non-extrusion), G1 is for linear moves (with optional extrusion).
-   */
-  g0(command: GCodeCommand, job: Job): void {
-    const { x, y, z, e, f } = command.params;
-
-    // discard zero length moves
-    if (x === undefined && y === undefined && z === undefined) {
-      if (e > 0) {
-        this.retractions++;
-      } else if (e < 0) {
-        this.deretractions++;
-      } else if (f !== undefined) {
-        this.feedrateChanges++;
-      } else {
-        this.others++;
-      }
-
-      return;
-    }
-
-    this.points++;
-
-    const { state } = job;
-    let currentPath = job.inprogressPath;
-    const pathType = e > 0 ? PathType.Extrusion : PathType.Travel;
-
-    if (currentPath === undefined || currentPath.travelType !== pathType) {
-      currentPath = this.breakPath(job, pathType);
-    }
-
-    if (e > 0) {
-      this.extrusionDistance += e;
-    }
-
-    // e is omitted bc currently we're assuming relative extrusion distances
-    // see also https://github.com/xyz-tools/gcode-preview/issues/179
-    state.x = x ?? state.x;
-    state.y = y ?? state.y;
-    state.z = z ?? state.z;
-
-    const pos = resolvePosition(state);
-    currentPath.addPoint(pos.x, pos.y, pos.z);
-    if (pathType === PathType.Extrusion) {
-      job.boundingBox.update(pos.x, pos.y, pos.z);
-    }
-  }
-
-  g1 = this.g0;
-
-  /**
-   * Executes an arc move command (G2/G3)
-   * @param command - GCodeCommand containing arc parameters
-   * @param job - Job instance to update
-   * @remarks
-   * Handles both clockwise (G2) and counter-clockwise (G3) arc moves. Supports
-   * both I/J center offset and R radius modes. Calculates intermediate points
-   * along the arc and updates the job state accordingly.
-   * G2 is for clockwise arcs, G3 is for counter-clockwise arcs.
-   */
-  g2(command: GCodeCommand, job: Job): void {
-    const { x, y, z, e } = command.params;
-    let { i, j, r } = command.params;
-    // Set when the arc cannot be described at all, so only the endpoint is emitted.
-    let arcIsDegenerate = false;
-    const { state } = job;
-    // Starting position for the arc, with any un-homed axis assumed at the origin.
-    const from = resolvePosition(state);
-
-    const cw = command.gcode === 'g2';
-    let currentPath = job.inprogressPath;
-    // `e > 0`, matching g0/g1: a negative E is a retraction, i.e. a travel move with no
-    // material laid down. The looser `e ?` used to misclassify a retracting arc as
-    // Extrusion, so it rendered as deposited filament and stretched the bounding box.
-    const pathType = e > 0 ? PathType.Extrusion : PathType.Travel;
-
-    if (currentPath === undefined || currentPath.travelType !== pathType) {
-      currentPath = this.breakPath(job, pathType);
-    }
-
-    if (e > 0) {
-      this.extrusionDistance += e;
-    }
-
-    if (r) {
-      // in r mode a minimum radius will be applied if the distance can otherwise not be bridged
-      const deltaX = x - from.x; // assume abs mode
-      const deltaY = y - from.y;
-
-      // apply a minimal radius to bridge the distance
-      const minR = Math.sqrt(Math.pow(deltaX / 2, 2) + Math.pow(deltaY / 2, 2));
-      r = Math.max(r, minR);
-
-      const dSquared = Math.pow(deltaX, 2) + Math.pow(deltaY, 2);
-      const hSquared = Math.pow(r, 2) - dSquared / 4;
-
-      // R mode cannot describe a whole circle: with start == end the centre is
-      // undefined, and hSquared / dSquared divides by zero, which used to make i/j NaN
-      // and poison every derived value. Skip the arc and move straight to the endpoint.
-      // (hSquared < 0 needs no guard: r was just clamped to minR, so it cannot go
-      // negative.) This replaces the guard that was commented out here.
-      if (dSquared === 0) {
-        arcIsDegenerate = true;
-      } else {
-        let hDivD = Math.sqrt(hSquared / dSquared);
-
-        // Ref RRF DoArcMove for details
-        if ((cw && r < 0.0) || (!cw && r > 0.0)) {
-          hDivD = -hDivD;
-        }
-        i = deltaX / 2 + deltaY * hDivD;
-        j = deltaY / 2 - deltaX * hDivD;
-      }
-    }
-
-    const wholeCircle = from.x == x && from.y == y;
-    const centerX = from.x + i;
-    const centerY = from.y + j;
-
-    const arcRadius = Math.sqrt(i * i + j * j);
-    const arcCurrentAngle = Math.atan2(-j, -i);
-    const finalTheta = Math.atan2(y - centerY, x - centerX);
-
-    let totalArc;
-    if (wholeCircle) {
-      totalArc = 2 * Math.PI;
-    } else {
-      totalArc = cw ? arcCurrentAngle - finalTheta : finalTheta - arcCurrentAngle;
-      if (totalArc < 0.0) {
-        totalArc += 2 * Math.PI;
-      }
-    }
-    let totalSegments = (arcRadius * totalArc) / 0.5;
-    if (state.units == 'in') {
-      totalSegments *= 25;
-    }
-    if (totalSegments < 1) {
-      totalSegments = 1;
-    }
-    let arcAngleIncrement = totalArc / totalSegments;
-    arcAngleIncrement *= cw ? -1 : 1;
-
-    // target - current. This was the other way round, so a helical arc climbing
-    // Z1 -> Z3 walked its intermediate points down to Z-0.97 and only landed on Z3 at
-    // the endpoint, leaving a visible spike in the rendered path.
-    //
-    // `??` not `||`: Z0 is a legitimate target, and `|| state.z` turned a descent to
-    // Z0 into a flat arc at the old height. This matches the endpoint assignment
-    // below, and is only safe because the parser now drops non-finite params -- `||`
-    // was rejecting NaN here by accident, and `??` does not.
-    const zDist = (z ?? from.z) - from.z;
-    const zStep = zDist / totalSegments;
-
-    // get points for the arc
-    let px = from.x;
-    let py = from.y;
-    let pz = from.z;
-    // calculate segments
-    let currentAngle = arcCurrentAngle;
-
-    // A degenerate arc leaves totalSegments non-finite even though every param was
-    // finite: an R-mode whole circle divides by dSquared === 0, and offsets near
-    // Number.MAX_VALUE overflow arcRadius. NaN already skipped the loop (NaN - 1
-    // fails the condition), but Infinity ran until the vertex array hit its length
-    // limit and threw, aborting the whole job. Emit no intermediate points in either
-    // case and fall through to the endpoint below.
-    if (!arcIsDegenerate && Number.isFinite(totalSegments)) {
-      for (let moveIdx = 0; moveIdx < totalSegments - 1; moveIdx++) {
-        currentAngle += arcAngleIncrement;
-        px = centerX + arcRadius * Math.cos(currentAngle);
-        py = centerY + arcRadius * Math.sin(currentAngle);
-        pz += zStep;
-        currentPath.addPoint(px, py, pz);
-        if (pathType === PathType.Extrusion) {
-          job.boundingBox.update(px, py, pz);
-        }
-      }
-    }
-
-    // `??` not `||`: an arc ending on X0, Y0 or Z0 used to silently keep the previous
-    // coordinate. Safe now that the parser drops non-finite params -- `||` was also
-    // rejecting NaN here by accident, which `??` does not do.
-    state.x = x ?? state.x;
-    state.y = y ?? state.y;
-    state.z = z ?? state.z;
-
-    const pos = resolvePosition(state);
-    currentPath.addPoint(pos.x, pos.y, pos.z);
-    if (pathType === PathType.Extrusion) {
-      job.boundingBox.update(pos.x, pos.y, pos.z);
-    }
-  }
-
-  g3 = this.g2;
-
-  /**
-   * Executes a G20 command to set units to inches
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   */
-  g20(command: GCodeCommand, job: Job): void {
-    job.state.units = 'in';
-  }
-
-  /**
-   * Executes a G21 command to set units to millimeters
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   */
-  g21(command: GCodeCommand, job: Job): void {
-    job.state.units = 'mm';
-  }
-
-  /**
-   * Executes a G28 homing command
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Moves all axes to their home positions (0,0,0) and marks the state as homed,
-   * so the position is now known rather than assumed.
-   */
-  g28(command: GCodeCommand, job: Job): void {
-    job.state.x = 0;
-    job.state.y = 0;
-    job.state.z = 0;
-    job.state.isHomed = true;
-  }
-
-  /**
-   * Selects tool 0 (T0)
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Updates the job state to use tool 0. Tools are typically used for
-   * multi-extruder setups or different print heads.
-   */
-  t0(command: GCodeCommand, job: Job): void {
-    job.state.tool = 0;
-  }
-  /**
-   * Selects tool 1 (T1)
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Updates the job state to use tool 1. Tools are typically used for
-   * multi-extruder setups or different print heads.
-   */
-  t1(command: GCodeCommand, job: Job): void {
-    job.state.tool = 1;
-  }
-  /**
-   * Selects tool 2 (T2)
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Updates the job state to use tool 2. Tools are typically used for
-   * multi-extruder setups or different print heads.
-   */
-  t2(command: GCodeCommand, job: Job): void {
-    job.state.tool = 2;
-  }
-  /**
-   * Selects tool 3 (T3)
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Updates the job state to use tool 3. Tools are typically used for
-   * multi-extruder setups or different print heads.
-   */
-  t3(command: GCodeCommand, job: Job): void {
-    job.state.tool = 3;
-  }
-  /**
-   * Selects tool 4 (T4)
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Updates the job state to use tool 4. Tools are typically used for
-   * multi-extruder setups or different print heads.
-   */
-  t4(command: GCodeCommand, job: Job): void {
-    job.state.tool = 4;
-  }
-  /**
-   * Selects tool 5 (T5)
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Updates the job state to use tool 5. Tools are typically used for
-   * multi-extruder setups or different print heads.
-   */
-  t5(command: GCodeCommand, job: Job): void {
-    job.state.tool = 5;
-  }
-  /**
-   * Selects tool 6 (T6)
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Updates the job state to use tool 6. Tools are typically used for
-   * multi-extruder setups or different print heads.
-   */
-  t6(command: GCodeCommand, job: Job): void {
-    job.state.tool = 6;
-  }
-  /**
-   * Selects tool 7 (T7)
-   * @param command - GCodeCommand containing the command
-   * @param job - Job instance to update
-   * @remarks
-   * Updates the job state to use tool 7. Tools are typically used for
-   * multi-extruder setups or different print heads.
-   */
-  t7(command: GCodeCommand, job: Job): void {
-    job.state.tool = 7;
-  }
-
-  /**
-   * Creates a new path and sets it as the current in-progress path
-   * @param job - Job instance to update
-   * @param newType - Type of the new path
-   * @returns The newly created path
-   * @remarks
-   * This method is called when a path type change is detected (e.g. switching
-   * between travel and extrusion moves). It finalizes the current path and
-   * starts a new one of the specified type.
-   */
-  private breakPath(job: Job, newType: PathType): Path {
-    job.finishPath();
-    const currentPath = new Path(newType, 0.6, 0.2, job.state.tool);
-    const pos = resolvePosition(job.state);
-    currentPath.addPoint(pos.x, pos.y, pos.z);
-    job.inprogressPath = currentPath;
-    return currentPath;
-  }
-}
